refactor(actions): archive/ como directorio con barril
